+++ packages/frontend/src/components/composer/EmojiAndStickerPicker.tsx
@@ -195,6 +195,8 @@ export const EmojiAndStickerPicker = forwardRef<
     hideStickerPicker?: boolean
   }
 >((props, ref) => {
+  const tx = useTranslationFunction()
+
   const accountId = selectedAccountId()
   const { onEmojiSelect, chatId, setShowEmojiPicker, hideStickerPicker } = props
 
@@ -223,7 +225,7 @@ export const EmojiAndStickerPicker = forwardRef<
             onClick={() => setShowSticker(false)}
             isSelected={!showSticker}
           >
-            Emoji
+            {tx('emoji')}
           </EmojiOrStickerSelectorButton>
           <EmojiOrStickerSelectorButton
             id='emoji-sticker-picker-tab-sticker'
@@ -231,7 +233,7 @@ export const EmojiAndStickerPicker = forwardRef<
             onClick={() => setShowSticker(true)}
             isSelected={showSticker}
           >
-            Sticker
+            {tx('sticker')}
           </EmojiOrStickerSelectorButton>
         </div>
       )}